vyatta-cfg-dataplane: Clear dataplane ext counters
Fixed an issue with addtion of qstat causing 'bad stat' error

The qstats added to clear files are bypassed when checking
for known stats.

VRVDR-52076
Simplified regex

(cherry picked from commit 85ca78ffe20a286bb3abba5e8f3b89fb5d54e45e)

+++ lib/Vyatta/DataplaneStats.pm
@@ -105,7 +105,11 @@ sub get_dataplane_clear_stats {
         # sanity check: if unknown stats field is in the clear file,
         # or clear value is greater than current one then don't trust the file,
         # shouldn't need to worry about wrap with 64bit counters
-        if ( !defined( $ifstat->{$var} ) || $val > $ifstat->{$var} ) {
+        # skip qstats, they are in clear file for future use.
+
+        if (($var !~ 'q([0-9]|[1-9][0-9]+)_\w+') &&
+            (!defined( $ifstat->{$var} ) ||
+             ($val > $ifstat->{$var} ))) {
             print "bad stat [$var, $intf]\n";
             close($f);
             unlink $filename;
